refactor: split core.py into focused modules
Extract from the 2,221-line core.py:
- _version.py (~100 lines): S3 version discovery, pure functions
- _bootstrap.py (~150 lines): SedonaDB init, caches, Overture loading
- _geometry.py (~150 lines): geometry mixin (wkt/wkb/geojson)

Also fixes code quality issues:
- Remove dead chain guards (already handled by __getattribute__)
- Replace fragile unicode repr detection with .count() == 0
- Pass sedona explicitly to _seed_country_info
- Extract _group_candidates from _ambiguity_message (120→80 lines)
